Versjonerer opp API til v1
