+++ mulighetsrommet-api/src/main/kotlin/no/nav/mulighetsrommet/api/Application.kt
@@ -7,7 +7,12 @@ import io.ktor.server.engine.*
 import io.ktor.server.netty.*
 import io.ktor.server.routing.*
 import no.nav.mulighetsrommet.api.plugins.*
-import no.nav.mulighetsrommet.api.routes.*
+import no.nav.mulighetsrommet.api.routes.internalRoutes
+import no.nav.mulighetsrommet.api.routes.swaggerRoutes
+import no.nav.mulighetsrommet.api.routes.v1.arenaRoutes
+import no.nav.mulighetsrommet.api.routes.v1.innsatsgruppeRoutes
+import no.nav.mulighetsrommet.api.routes.v1.tiltaksgjennomforingRoutes
+import no.nav.mulighetsrommet.api.routes.v1.tiltakstypeRoutes
 import org.slf4j.LoggerFactory
 
 fun main() {
