+++ mulighetsrommet-arena-adapter/src/main/kotlin/no/nav/mulighetsrommet/arena/adapter/consumers/SakEndretConsumer.kt
@@ -10,7 +10,7 @@ import org.slf4j.LoggerFactory
 class SakEndretConsumer(private val client: MulighetsrommetApiClient) {
 
     private val logger = LoggerFactory.getLogger(SakEndretConsumer::class.java)
-    private var resourceUri = "/api/arena/sak"
+    private var resourceUri = "/api/v1/arena/sak"
 
     fun process(payload: JsonElement) {
         if (isInsertArenaOperation(payload.jsonObject)) handleInsert(payload.jsonObject) else handleUpdate(payload.jsonObject)
