+++ mulighetsrommet-arena-adapter/src/main/kotlin/no/nav/mulighetsrommet/arena/adapter/consumers/TiltakEndretConsumer.kt
@@ -11,7 +11,7 @@ import org.slf4j.LoggerFactory
 class TiltakEndretConsumer(private val client: MulighetsrommetApiClient) {
 
     private val logger = LoggerFactory.getLogger(TiltakEndretConsumer::class.java)
-    private var resourceUri = "/api/arena/tiltakstyper"
+    private var resourceUri = "/api/v1/arena/tiltakstyper"
 
     fun process(payload: JsonElement) {
         if (isInsertArenaOperation(payload.jsonObject)) handleInsert(payload.jsonObject) else handleUpdate(payload.jsonObject)
