+++ mulighetsrommet-arena-adapter/wiremock/mappings/mr-tiltakstype-post.json
@@ -1,7 +1,7 @@
 {
   "request": {
     "method": "POST",
-    "url": "/api/tiltakstyper"
+    "url": "/api/v1/tiltakstyper"
   },
   "response": {
     "status": 201,
